% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_tagset}
\alias{get_tagset}
\title{Get tags of selected types from an XML file}
\usage{
get_tagset(x, nodes = c("hedge"))
}
\arguments{
\item{x}{filename}

\item{nodes}{vector of tag names to extract (default: just 'hedge')}
}
\value{
a \code{data.frame} with headings:
\itemize{
 \item \code{file} name of file
 \item \code{id} tag identifier, e.g. hedge
 \item \code{start} letter index of annotation start
 \item \code{end} letter index of annotation end
 \item \code{text} text of annotation
 \item \code{type} value of 'type' attribute, if there is one
 \item \code{here} value of 'here' attribute, if there is one
}
}
\description{
Get tags of selected types from an XML file
}

