% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{match_nodes}
\alias{match_nodes}
\title{Match different nodes based on start and endpoints}
\usage{
match_nodes(tagset, match_x, match_y)
}
\arguments{
\item{tagset}{Data-frame returned by \code{get_tagset} or \code{get_tagsets}}

\item{match_x}{Tag value of node to be matched on}

\item{match_y}{Tag value to be matched}
}
\value{
a \code{data.frame} of  similar structure as the return of \code{get_tagset} with values of nodes specified in \code{match_y} matched to the corresponding rows of nodes specified in \code{match_x} based on the start-end intervals.
}
\description{
Match different nodes based on start and endpoints
}

