% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlAnnotate.R
\docType{package}
\name{xmlAnnotate}
\alias{xmlAnnotate}
\alias{xmlAnnotate-package}
\title{xmlAnnotate: A package for extracting annotations.}
\description{
the xmlAnnotate package contains three functions: \code{get_tagset} for extracting
tagset information from an XML output file in a rather particular format,
\code{get_tagsets} which applies this to lots of such files and concatenates the
results, and \code{xml_grab}, which does, umm, something similar. You should probably
ask Jakob about that.
}
\details{
There's a vignette, showing the first two functions at work.
}

